requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
pillow
tqdm
kagglehub
opencv-python
albumentations
onnxruntime
